Workaround HALSimWS SimDevice deadlocks.
The critical change here is calling waitForUserToStart() *before* creating the robot. That gives the controller's HALSimWS client a chance to connect before the SimDevices are created. Before that change, I'd get a deadlock at ~1/7 of the time. With that change, I've run over 137 times in a row without a deadlock.